Verificar accesos mediante Spatie Controladores

+++ app/Http/Controllers/Dashboard/CategoryController.php
@@ -15,6 +15,11 @@ class CategoryController extends Controller
      */
     public function index()
     {
+        // Verificamos los accesos mediante Spatie
+        if (!auth()->user()->hasPermissionTo('editor.category.index')) {
+            return abort(403);
+        }
+
         $categories = Category::paginate(2);
         return view('dashboard/category/index',compact('categories'));
     }
@@ -24,6 +29,10 @@ class CategoryController extends Controller
      */
     public function create()
     {
+        if (!auth()->user()->hasPermissionTo('editor.category.create')) {
+            return abort(403);
+        }
+
         $category = new Category();
 
         return view('dashboard/category/create',compact('category'));
@@ -34,6 +43,10 @@ class CategoryController extends Controller
      */
     public function store(StoreRequest $request)
     {
+        if (!auth()->user()->hasPermissionTo('editor.category.create')) {
+            return abort(403);
+        }
+
         Category::create($request->validated());
 
         // Mensaje de tipo Flash para indicar al usuario que la accion se realizo
@@ -62,6 +75,10 @@ class CategoryController extends Controller
      */
     public function show(Category $category)
     {
+        if (!auth()->user()->hasPermissionTo('editor.category.index')) {
+            return abort(403);
+        }
+
         return view('dashboard/category/show', compact('category'));
     }
 
@@ -70,6 +87,10 @@ class CategoryController extends Controller
      */
     public function edit(Category $category)
     {
+        if (!auth()->user()->hasPermissionTo('editor.category.update')) {
+            return abort(403);
+        }
+
         return view('dashboard/category/edit', compact('category'));
     }
 
@@ -78,6 +99,10 @@ class CategoryController extends Controller
      */
     public function update(PutRequest $request, Category $category)
     {
+        if (!auth()->user()->hasPermissionTo('editor.category.update')) {
+            return abort(403);
+        }
+
         $category->update($request->validated());
 
         return to_route('category.index')->with('status', 'Category Updated');
@@ -88,6 +113,10 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
+        if (!auth()->user()->hasPermissionTo('editor.category.destroy')) {
+            return abort(403);
+        }
+
         $category->delete();
 
         return to_route('category.index')->with('status', 'Category Deleted');
